Move device-map prune and rank policy into Rust (ABI 9); make ApiHttpServer buildable on Android
BleScanEngine's four-level Java sort comparator and its "drop when stale"
rule were the last device-map policy still implemented in Java. Both now
run in bleradar-jni: deviceShouldPrune(freshness) and deviceRankKey(
freshness, lastSeenUptimeMs, confidence, rssi) pack the ranking into one
non-negative long so snapshot() sorts by a key sampled once per device,
which also removes a latent TimSort contract violation (the old comparator
read volatile Blip fields mid-sort). Freshness refresh and pruning are one
pass now instead of two. ABI bumped 8 -> 9; contract gate 27 <-> 27.

ApiHttpServer used com.sun.net.httpserver, a JDK-only package absent from
android.jar, so it could never build for Android; it is now a minimal
HTTP/1.1 responder over java.net.ServerSocket. last_seen_ago_ms subtracted
an uptime timestamp from wall-clock epoch; it is now measured on
SystemClock.uptimeMillis(). UpdateManager.computeRetryBackoff passed a
long attempt into the int native and failed javac against android.jar.

Regression locks: five jni_bridge tests (prune ordinal contract, rank
tiers, clamping, non-negativity, 100k random pairs vs a reference
comparator), the export campaign now covers both new exports, and
verify-jni-live asserts both natives across a real JVM (abi=9,
linked-natives=27).

+++ android/app/src/main/java/com/hse/bleradar/UpdateManager.java
@@ -154,7 +154,7 @@ public final class UpdateManager {
      * @param maxDelaySecs cap on the computed delay (e.g. 86400 for 24h)
      * @return seconds to wait before the next retry
      */
-    public long computeRetryBackoff(long attemptNumber, long baseDelaySeconds, long maxDelaySecs) {
+    public long computeRetryBackoff(int attemptNumber, long baseDelaySeconds, long maxDelaySecs) {
         if (!NativeRadar.isAvailable()) {
             return baseDelaySeconds;
         }
